// ASEDINFO_FRONTEND/src/app/main/pages/compartidos/interfaces/IusuarioRolAplicacion.ts
export interface IusuarioRolAplicacion {
    codigo?: number;
    codRolAplicacion?: number;
    rolAplicacion?: any;
    codUsuario?: number;
    usuario?: any;
    fechaInicial?: string;
    fechaFinal?: string;
    estado?: string;

    codigoRolAplicacionRecurso?: number;
    rolAplicacionRecurso: any;

}